+++ doc/changelog.rst
@@ -5,6 +5,7 @@ Version 0.9.5
 -------------
 
 * Add ``--python-names`` option to the :ref:`Runner CLI <runner_cli>`
+* ``system show`` command now checks if the system is ready for benchmarking
 
 Version 0.9.4 (2017-03-01)
 --------------------------